FlexiTradeBot/requirements.txt
python-binance==1.0.19
pandas==2.1.0
numpy==1.24.3
python-dotenv==1.0.0
websocket-client==1.6.1
ccxt==4.1.13
colorama==0.4.6
